# Update to 1.18.2, update Gradle wrapper, fix support for various mods