… gradle/wrapper/gradle-wrapper.jar
… gradlew
… src/main/java/me/modmuss50/optifabric/compat/zoomify/mixin/GameRendererMixin.java
… src/main/java/me/modmuss50/optifabric/mixin/CustomColoursMixin.java
… src/main/java/me/modmuss50/optifabric/mixin/ParticleManagerMixin.java
… src/main/java/me/modmuss50/optifabric/mod/OptifabricSetup.java
… src/main/java/me/modmuss50/optifabric/patcher/fixes/KeyboardFix.java
… src/main/resources/fabric.mod.json
… src/main/resources/optifabric.compat.apoli-new.mixins.json
… src/main/resources/optifabric.compat.apoli-newer.mixins.json
… src/main/resources/optifabric.compat.architectury-AB.newerer-mixins.json
… src/main/resources/optifabric.compat.carpet.extra-new-mixins.json
… src/main/resources/optifabric.compat.replaymod.newer-mixins.json
… src/main/resources/optifabric.compat.zoomify.mixins.json
… src/main/resources/optifabric.optifine.mixins.json
… src/shim/java/net/minecraft/class_5636.java