## For anyone skimming this thread: There is an experimental jar available [in this comment in my pull request](https://github.com/Chocohead/OptiFabric/pull/630#issuecomment-1075090945). Using Architectury? - you will have to download the [version without blacklist](https://github.com/Chocohead/OptiFabric/pull/630#issuecomment-1066655207) from this thread

**Edit:** Further progress, the game now works but mining blocks causes a crash.
**Edit 2:** There is now a working jar available at the end of this thread

This commit on this pull request simply changes `fabric.mod.json` to include `"1.18.2"`. Doing this causes an infinite boot loop: [log.txt](https://github.com/Chocohead/OptiFabric/files/8238450/log.txt)
```
java.util.concurrent.CompletionException: java.lang.RuntimeException: Mixin transformation of net.optifine.CustomColors failed
	at java.util.concurrent.CompletableFuture.encodeThrowable(CompletableFuture.java:315) ~[?:?]
	at java.util.concurrent.CompletableFuture.completeThrowable(CompletableFuture.java:320) ~[?:?]
	at java.util.concurrent.CompletableFuture$UniRun.tryFire(CompletableFuture.java:791) ~[?:?]
	at java.util.concurrent.CompletableFuture$Completion.run(CompletableFuture.java:482) ~[?:?]
	at net.minecraft.class_4014.method_18365(class_4014.java:69) ~[client-intermediary.jar:?]
	at net.minecraft.class_1255.execute(class_1255.java:118) ~[client-intermediary.jar:?]
	at net.minecraft.class_4014.method_18367(class_4014.java:68) ~[client-intermediary.jar:?]
[...]
Caused by: org.spongepowered.asm.mixin.transformer.throwables.MixinTransformerError: An unexpected critical error was encountered
	at org.spongepowered.asm.mixin.transformer.MixinProcessor.applyMixins(MixinProcessor.java:392) ~[sponge-mixin-0.11.2+mixin.0.8.5.jar:0.11.2+mixin.0.8.5]
	at org.spongepowered.asm.mixin.transformer.MixinTransformer.transformClass(MixinTransformer.java:234) ~[sponge-mixin-0.11.2+mixin.0.8.5.jar:0.11.2+mixin.0.8.5]
	at org.spongepowered.asm.mixin.transformer.MixinTransformer.transformClassBytes(MixinTransformer.java:202) ~[sponge-mixin-0.11.2+mixin.0.8.5.jar:0.11.2+mixin.0.8.5]
	at net.fabricmc.loader.impl.launch.knot.KnotClassDelegate.getPostMixinClassByteArray(KnotClassDelegate.java:247) ~[fabric-loader-0.13.3.jar:?]
	at net.fabricmc.loader.impl.launch.knot.KnotClassDelegate.tryLoadClass(KnotClassDelegate.java:150) ~[fabric-loader-0.13.3.jar:?]
[...]
Caused by: org.spongepowered.asm.mixin.injection.throwables.InjectionError: Critical injection failure: Callback method skip(Lnet/minecraft/class_777;Lnet/minecraft/class_2680;Lnet/minecraft/class_1920;Lnet/minecraft/class_2338;Ljava/lang/Object;Lorg/spongepowered/asm/mixin/injection/callback/CallbackInfoReturnable;)V in optifabric.optifine.mixins.json:CustomColoursMixin from mod (unknown) failed injection check, (0/1) succeeded. Scanned 1 target(s). Using refmap optifabric-refmap.json
	at org.spongepowered.asm.mixin.injection.struct.InjectionInfo.postInject(InjectionInfo.java:468) ~[sponge-mixin-0.11.2+mixin.0.8.5.jar:0.11.2+mixin.0.8.5]
	at org.spongepowered.asm.mixin.transformer.MixinTargetContext.applyInjections(MixinTargetContext.java:1385) ~[sponge-mixin-0.11.2+mixin.0.8.5.jar:0.11.2+mixin.0.8.5]
	at org.spongepowered.asm.mixin.transformer.MixinApplicatorStandard.applyInjections(MixinApplicatorStandard.java:1052) ~[sponge-mixin-0.11.2+mixin.0.8.5.jar:0.11.2+mixin.0.8.5]
```

I've also changed the build process since the build sometimes broke with a duplicate resource error. Not sure if it was originally like that for a reason but now it matches what the `fabric-example-mod` repo is like.

```diff
--- a/build.gradle
+++ b/build.gradle
@@ -81,14 +81,9 @@ task unzip(type: Copy) {
 processResources {
 	inputs.property "version", project.version
 
-	from(sourceSets.main.resources.srcDirs) {
-		include "fabric.mod.json"
+	filesMatching("fabric.mod.json") {
 		expand "version": project.version
 	}
-
-	from(sourceSets.main.resources.srcDirs) {
-		exclude "fabric.mod.json"
-	}
 }
 
 // ensure that the encoding is set to UTF-8, no matter what the system default is
```

Related issues:
- #618 
- #619
- #622
- #625
- #626
- #627