@MrFawkes1337 2022-03-12
Theres more changes involved than just simply adding the version. A good place to start would be to diff the changes made between Optifine HD U H5 Pre-3 or pre-4 with the next version up, as it doesn't work due to the error you describe, then you have a basis to scroll through commits to find where you need to start applying a fix. Im looking to do just that. You'll also need to update loader support, etc.